+++ readium/navigators/web/fixedlayout/src/main/kotlin/org/readium/navigator/web/fixedlayout/FixedWebRenditionState.kt
@@ -62,6 +62,7 @@ import org.readium.r2.navigator.preferences.Axis
 import org.readium.r2.navigator.preferences.Fit
 import org.readium.r2.shared.ExperimentalReadiumApi
 import org.readium.r2.shared.InternalReadiumApi
+import org.readium.r2.shared.util.AbsoluteUrl
 import org.readium.r2.shared.util.RelativeUrl
 import org.readium.r2.shared.util.Url
 import org.readium.r2.shared.util.mediatype.MediaType
@@ -155,12 +156,19 @@ public class FixedWebRenditionState internal constructor(
             application = application,
             container = publication.container,
             mediaTypes = publication.mediaTypes,
+            baseUrl = publication.baseUrl,
             errorPage = RelativeUrl("readium/navigator/web/internals/error.xhtml")!!,
             htmlInjector = htmlInjector,
             servedAssets = configuration.servedAssets + listOf("readium/.*"),
             onResourceLoadFailed = { _, _ -> }
         )
     }
+
+    internal fun servedUrlToHref(url: AbsoluteUrl) =
+        webViewServer.servedUrlToHref(url)
+
+    internal fun hrefToServedUrl(href: Url) =
+        webViewServer.hrefToServedUrl(href)
 
     internal val webViewClient: WebViewClient =
         WebViewClient(webViewServer)