+++ readium/navigators/web/reflowable/src/main/kotlin/org/readium/navigator/web/reflowable/resource/ReflowableResource.kt
@@ -71,14 +71,13 @@ import org.readium.navigator.web.reflowable.ReflowableWebDecorationTextQuoteLoca
 import org.readium.navigator.web.reflowable.css.ReadiumCssInjector
 import org.readium.r2.shared.ExperimentalReadiumApi
 import org.readium.r2.shared.util.AbsoluteUrl
-import org.readium.r2.shared.util.Url
 import timber.log.Timber
 
 @SuppressLint("SetJavaScriptEnabled", "ClickableViewAccessibility")
 @Composable
 internal fun ReflowableResource(
     resourceState: ReflowableResourceState,
-    publicationBaseUrl: AbsoluteUrl,
+    servedUrl: AbsoluteUrl,
     webViewClient: WebViewClient,
     backgroundColor: Color,
     padding: AbsolutePaddingValues,
@@ -91,7 +90,7 @@ internal fun ReflowableResource(
     actionModeCallback: ActionMode.Callback?,
     onSelectionApiChanged: (ReflowableSelectionApi?) -> Unit,
     onTap: (TapEvent) -> Unit,
-    onLinkActivated: (Url, String) -> Unit,
+    onLinkActivated: (AbsoluteUrl, String) -> Unit,
     onDecorationActivated: (DecorationListener.OnActivatedEvent<ReflowableWebDecorationLocation>) -> Unit,
     onLocationChange: () -> Unit,
     onDocumentResized: () -> Unit,
@@ -101,7 +100,7 @@ internal fun ReflowableResource(
         propagateMinConstraints = true
     ) {
         val webViewState = rememberWebViewState<RelaxedWebView>(
-            url = publicationBaseUrl.resolve(resourceState.href).toString()
+            url = servedUrl.toString()
         )
 
         var documentStateApi by remember(webViewState.webView) {
@@ -209,6 +208,7 @@ internal fun ReflowableResource(
                                             direction = layoutDirection
                                         )
                                         onLocationChange()
+                                        showPlaceholder.value = false
                                     }
                                     null -> {
                                         scrollController.moveToProgression(
@@ -223,6 +223,7 @@ internal fun ReflowableResource(
                                             direction = layoutDirection
                                         )
                                         onLocationChange()
+                                        showPlaceholder.value = false
                                     }
                                 }
 
@@ -230,7 +231,6 @@ internal fun ReflowableResource(
                                     resourceState.updateProgression(orientation, layoutDirection)
                                     onLocationChange()
                                 }
-                                showPlaceholder.value = false
                             }
                         },
                         onDocumentResizedDelegate = {
@@ -255,6 +255,8 @@ internal fun ReflowableResource(
                         resourceState.pendingLocation
                     }.onEach { pendingLocation ->
                         pendingLocation?.let {
+                            showPlaceholder.value = true
+
                             when (pendingLocation) {
                                 is ReflowableResourceLocation.Progression -> {
                                     scrollController.moveToProgression(
@@ -303,6 +305,7 @@ internal fun ReflowableResource(
                                 direction = layoutDirection
                             )
                             onLocationChange()
+                            showPlaceholder.value = false
                         }
                     }.launchIn(this)
                 }
@@ -328,9 +331,7 @@ internal fun ReflowableResource(
                             val shiftedOffset = offset + paddingShift
                             onTap(TapEvent(shiftedOffset))
                         },
-                        onLinkActivatedDelegate = { href, outerHtml ->
-                            onLinkActivated(publicationBaseUrl.relativize(href), outerHtml)
-                        },
+                        onLinkActivatedDelegate = onLinkActivated,
                         onDecorationActivatedDelegate = { id, group, rect, offset ->
                             val decoration = decorations.value[group]?.firstOrNull { it.id.value == id }
                                 ?: return@DelegatingGesturesListener